Merging
Car splits off, then stops at a point.
When the cars drive by, they get highlighted red. If you press space while
one is red, the train will split there and you will be inserted.
Works fairly reliably. But is fairly fragile. If I move any of the node
(which there are many), it probably breaks. And I'll probably have to move
stuff when I make the train a different size/shape. The cars will probably
become rectangles instead of cubes so they will need more space between
one another.

Stopping and starting is a bit more stable. Can stop 2 cars. It only works
for 2 atm.
Starting is better. The speed boost propagates down the line so they all
get the boost. But cars past 2 won't be stopped in the right spot so it
will still look weird.

There has got to be a better way to do all of this. I just don't know what
it is.