Art and UI
Added a model for the train cars
The red tint indicator doesn't work for the models. I think I could create
my own materials and drag them in and it might work.
I might jsut do something other than a colour tint for an indicator.

Added UI for the speed indicator. Hooked it up to the train so it shows
the current speed.

Added a texture for crates.