Splitter Track, UI Art
Made a splitter track. Press R to switch the direction. Won't switch if
there are cars on it. Unless the cars on it belong to a different train
than the cars behind it.

Fixed bugs in ejector rail stuff. Wouldn't inject if there was nothign
waiting (inject at the end of the train).

UI indicators
UI for ejector track and the insert.
Kinda ugly, but functional. Should they just be white instead of coloured?

Rail Art
Made a rail texture
All the rail points make lineRenders with that tex from their point to the
next. Looks ok.

Train Model
Added a model for the engine. Looks decent.