Town 2, Crate Spawner, Claw Arm
Made the second town (town A) and the second slide rail for splitting and
inserting train cars

Made a crate spawner. It spawns a specific crate or random every x
seconds, but only if the old crate isn't still there.

Gave an arm to the claws.
It's a line renderer.

TownTollSwitch
Changed how going to town works. It now forces the last car into town. It
also takes any lone cars just rolling by. This might be unintended, but I
might leave it. Kinda bad design to have it so close to the merge spot and
it doesn't have enough speed to reach the train if you're going fast.
Many ways to solve this. Which is best?

Bugs
Fixed a bug with inserting trains. It tried to split the train in front,
but didn't check to see if it still had that front train. So it was
splitting the engine halfway across the map that was pulling something
else.